+++ dolphinscheduler-dao/src/main/java/org/apache/dolphinscheduler/dao/upgrade/ScheduleDao.java
@@ -17,8 +17,6 @@
 
 package org.apache.dolphinscheduler.dao.upgrade;
 
-import org.apache.dolphinscheduler.common.utils.ConnectionUtils;
-
 import java.sql.Connection;
 import java.sql.PreparedStatement;
 import java.sql.ResultSet;
@@ -42,11 +40,9 @@ public class ScheduleDao {
     public Map<Integer, Long> queryAllSchedule(Connection conn) {
         Map<Integer, Long> scheduleMap = new HashMap<>();
         String sql = "SELECT id,process_definition_code FROM t_ds_schedules";
-        ResultSet rs = null;
-        PreparedStatement pstmt = null;
-        try {
-            pstmt = conn.prepareStatement(sql);
-            rs = pstmt.executeQuery();
+        try (
+                PreparedStatement pstmt = conn.prepareStatement(sql);
+                ResultSet rs = pstmt.executeQuery()) {
             while (rs.next()) {
                 Integer id = rs.getInt(1);
                 long processDefinitionCode = rs.getLong(2);
@@ -55,8 +51,6 @@ public class ScheduleDao {
         } catch (Exception e) {
             logger.error(e.getMessage(), e);
             throw new RuntimeException("sql: " + sql, e);
-        } finally {
-            ConnectionUtils.releaseResource(rs, pstmt, conn);
         }
         return scheduleMap;
     }
@@ -92,8 +86,6 @@ public class ScheduleDao {
         } catch (Exception e) {
             logger.error(e.getMessage(), e);
             throw new RuntimeException("sql: " + sql, e);
-        } finally {
-            ConnectionUtils.releaseResource(conn);
         }
     }
 }
