+++ dolphinscheduler-tools/src/main/java/org/apache/dolphinscheduler/tools/datasource/upgrader/v130/V130DolphinSchedulerUpgrader.java
@@ -1,0 +1,97 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one or more
+ * contributor license agreements.  See the NOTICE file distributed with
+ * this work for additional information regarding copyright ownership.
+ * The ASF licenses this file to You under the Apache License, Version 2.0
+ * (the "License"); you may not use this file except in compliance with
+ * the License.  You may obtain a copy of the License at
+ *
+ *    http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.apache.dolphinscheduler.tools.datasource.upgrader.v130;
+
+import org.apache.dolphinscheduler.common.utils.JSONUtils;
+import org.apache.dolphinscheduler.dao.upgrade.ProcessDefinitionDao;
+import org.apache.dolphinscheduler.dao.upgrade.WorkerGroupDao;
+import org.apache.dolphinscheduler.tools.datasource.upgrader.DolphinSchedulerUpgrader;
+import org.apache.dolphinscheduler.tools.datasource.upgrader.DolphinSchedulerVersion;
+
+import java.sql.Connection;
+import java.util.HashMap;
+import java.util.Map;
+
+import javax.sql.DataSource;
+
+import lombok.extern.slf4j.Slf4j;
+
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Component;
+
+import com.fasterxml.jackson.databind.node.ArrayNode;
+import com.fasterxml.jackson.databind.node.ObjectNode;
+
+@Slf4j
+@Component
+public class V130DolphinSchedulerUpgrader implements DolphinSchedulerUpgrader {
+
+    @Autowired
+    private DataSource dataSource;
+
+    @Override
+    public void doUpgrade() {
+        updateProcessDefinitionJsonWorkerGroup();
+    }
+
+    private void updateProcessDefinitionJsonWorkerGroup() {
+        WorkerGroupDao workerGroupDao = new WorkerGroupDao();
+        ProcessDefinitionDao processDefinitionDao = new ProcessDefinitionDao();
+        Map<Integer, String> replaceProcessDefinitionMap = new HashMap<>();
+        try (Connection connection = dataSource.getConnection()) {
+            Map<Integer, String> oldWorkerGroupMap = workerGroupDao.queryAllOldWorkerGroup(connection);
+            Map<Integer, String> processDefinitionJsonMap =
+                    processDefinitionDao.queryAllProcessDefinition(connection);
+
+            for (Map.Entry<Integer, String> entry : processDefinitionJsonMap.entrySet()) {
+                ObjectNode jsonObject = JSONUtils.parseObject(entry.getValue());
+                ArrayNode tasks = JSONUtils.parseArray(jsonObject.get("tasks").toString());
+
+                for (int i = 0; i < tasks.size(); i++) {
+                    ObjectNode task = (ObjectNode) tasks.path(i);
+                    ObjectNode workerGroupNode = (ObjectNode) task.path("workerGroupId");
+                    int workerGroupId = -1;
+                    if (workerGroupNode != null && workerGroupNode.canConvertToInt()) {
+                        workerGroupId = workerGroupNode.asInt(-1);
+                    }
+                    if (workerGroupId == -1) {
+                        task.put("workerGroup", "default");
+                    } else {
+                        task.put("workerGroup", oldWorkerGroupMap.get(workerGroupId));
+                    }
+                }
+
+                jsonObject.remove("task");
+
+                jsonObject.put("tasks", tasks);
+
+                replaceProcessDefinitionMap.put(entry.getKey(), jsonObject.toString());
+            }
+            if (replaceProcessDefinitionMap.size() > 0) {
+                processDefinitionDao.updateProcessDefinitionJson(connection, replaceProcessDefinitionMap);
+            }
+        } catch (Exception e) {
+            log.error("update process definition json workergroup error", e);
+        }
+    }
+
+    @Override
+    public DolphinSchedulerVersion getCurrentVersion() {
+        return DolphinSchedulerVersion.V1_3_0;
+    }
+}
